hw4/requirements.txt
requests
beautifulsoup4
elasticsearch
networkx
pandas
PyPDF2
python-docx
scikit-learn
numpy
pypinyin
jieba
openpyxl
aiohttp